moved Payments tab handling to new class
