+++ src/Tickets/Settings.php
@@ -8,6 +8,8 @@
 
 namespace TEC\Tickets;
 
+use TEC\Tickets\Commerce\Payments_Tab;
+
 /**
  * Class Settings.
  *
@@ -24,4 +26,24 @@ class Settings {
 	 * @var string
 	 */
 	public static $tickets_commerce_enabled = 'tickets_commerce_enabled';
+
+	/**
+	 * Force enable TicketsCommerce for new users.
+	 *
+	 * @since TBD
+	 *
+	 * @param bool $enabled
+	 *
+	 * @return bool|mixed
+	 */
+	public static function force_enable_tickets_commerce( $enabled ) {
+
+		return true;
+		// If the tickets_commerce setting was saved before, then maintain the setting value.
+		if ( '' != tribe_get_option( Payments_Tab::$option_enable ) ) {
+			return $enabled;
+		}
+
+		return tribe_installed_after( 'Tribe__Tickets__Main', '5.1.10' );
+	}
 }
